moving tests to correct place
