+++ tests/ui_tests/test_ui_file_writer_ctrl_window.py
@@ -1,15 +1,35 @@
 import pytest
+from PySide2.QtCore import QSettings
 from PySide2.QtGui import QStandardItemModel
 from mock import Mock
 from streaming_data_types import run_start_pl72
-from nexus_constructor.file_writer_ctrl_window import FileWriterCtrl, File, FileWriter
 from nexus_constructor.validators import BrokerAndTopicValidator
+from nexus_constructor.file_writer_ctrl_window import (
+    FileWriterCtrl,
+    FileWriterSettings,
+    extract_bool_from_qsettings,
+    File,
+    FileWriter,
+)
+
+
+@pytest.fixture()
+def settings():
+    settings = QSettings("testing", "NCui_tests")
+    yield settings
+    settings.setValue(FileWriterSettings.STATUS_BROKER_ADDR, "")
+    settings.setValue(FileWriterSettings.COMMAND_BROKER_ADDR, "")
+    settings.setValue(FileWriterSettings.FILE_NAME, "")
+    settings.setValue(FileWriterSettings.USE_START_TIME, False)
+    settings.setValue(FileWriterSettings.USE_STOP_TIME, False)
+    settings.setValue(FileWriterSettings.FILE_BROKER_ADDR, "")
+    del settings
 
 
 def test_UI_GIVEN_nothing_WHEN_creating_filewriter_control_window_THEN_broker_field_defaults_are_set_correctly(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
 
     assert not window.command_broker_edit.text()
@@ -22,9 +42,9 @@ def test_UI_GIVEN_nothing_WHEN_creating_filewriter_control_window_THEN_broker_fi
 
 
 def test_UI_GIVEN_nothing_WHEN_creating_filewriter_control_window_THEN_broker_validators_are_set_correctly(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
 
     assert isinstance(window.status_broker_edit.validator(), BrokerAndTopicValidator)
@@ -49,9 +69,9 @@ def test_UI_GIVEN_time_string_WHEN_setting_time_THEN_last_time_is_stored(
 
 
 def test_UI_GIVEN_no_files_WHEN_stop_file_writing_is_clicked_THEN_button_is_disabled(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
     window.files_list.selectedIndexes = lambda: []
 
@@ -61,9 +81,9 @@ def test_UI_GIVEN_no_files_WHEN_stop_file_writing_is_clicked_THEN_button_is_disa
 
 
 def test_UI_GIVEN_files_WHEN_stop_file_writing_is_clicked_THEN_button_is_enabled(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
     window.files_list.selectedIndexes = lambda: [
         1,
@@ -77,10 +97,10 @@ def test_UI_GIVEN_files_WHEN_stop_file_writing_is_clicked_THEN_button_is_enabled
 
 
 def test_UI_GIVEN_valid_command_WHEN_sending_command_THEN_command_producer_sends_command(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
 
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
     window.command_producer = Mock()
 
@@ -108,9 +128,9 @@ def test_UI_GIVEN_valid_command_WHEN_sending_command_THEN_command_producer_sends
 
 
 def test_UI_GIVEN_no_status_consumer_and_no_command_producer_WHEN_checking_status_connection_THEN_both_leds_are_turned_off(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
     window.status_consumer = None
     window.command_producer = None
@@ -122,9 +142,9 @@ def test_UI_GIVEN_no_status_consumer_and_no_command_producer_WHEN_checking_statu
 
 
 def test_UI_GIVEN_status_consumer_but_no_command_producer_WHEN_checking_status_connection_THEN_status_led_is_turned_on(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
     window.command_producer = None
     window.status_consumer = Mock()
@@ -137,9 +157,9 @@ def test_UI_GIVEN_status_consumer_but_no_command_producer_WHEN_checking_status_c
 
 
 def test_UI_GIVEN_command_producer_WHEN_checking_connection_status_THEN_command_led_is_turned_on(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
     window.command_producer = Mock()
     window.status_consumer = None
@@ -157,9 +177,9 @@ class DummyInterface:
 
 @pytest.mark.skip(reason="qtbot interferes with other tests")
 def test_UI_GIVEN_invalid_broker_WHEN_status_broker_timer_callback_is_called_THEN_nothing_happens(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
     window.status_consumer = None
     window.status_broker_edit.setText("invalid")
@@ -170,9 +190,9 @@ def test_UI_GIVEN_invalid_broker_WHEN_status_broker_timer_callback_is_called_THE
 
 @pytest.mark.skip(reason="qtbot interferes with other tests")
 def test_UI_GIVEN_invalid_broker_WHEN_command_broker_timer_callback_is_called_THEN_nothing_happens(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
     window.command_producer = None
     window.command_broker_edit.setText("invalid")
@@ -183,9 +203,9 @@ def test_UI_GIVEN_invalid_broker_WHEN_command_broker_timer_callback_is_called_TH
 
 @pytest.mark.skip(reason="qtbot interferes with other tests")
 def test_UI_GIVEN_valid_broker_WHEN_command_broker_timer_callback_is_called_THEN_producer_is_created(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
     window.command_broker_change_timer.stop()
     window.status_broker_change_timer.stop()
@@ -199,9 +219,9 @@ def test_UI_GIVEN_valid_broker_WHEN_command_broker_timer_callback_is_called_THEN
 
 @pytest.mark.skip(reason="qtbot interferes with other tests")
 def test_UI_GIVEN_valid_broker_WHEN_status_broker_timer_callback_is_called_THEN_consumer_is_created(
-    qtbot, instrument
+    qtbot, instrument, settings
 ):
-    window = FileWriterCtrl(instrument)
+    window = FileWriterCtrl(instrument, settings)
     qtbot.addWidget(window)
     window.command_broker_change_timer.stop()
     window.status_broker_change_timer.stop()
@@ -211,3 +231,78 @@ def test_UI_GIVEN_valid_broker_WHEN_status_broker_timer_callback_is_called_THEN_
 
     window.status_broker_timer_changed(DummyInterface)
     assert isinstance(window.status_consumer, DummyInterface)
+
+
+def test_UI_settings_are_saved_when_store_settings_is_called(
+    qtbot, instrument, settings
+):
+    window = FileWriterCtrl(instrument, settings)
+    qtbot.addWidget(window)
+
+    command_broker = "broker:9092/topic1"
+    window.command_broker_edit.setText(command_broker)
+
+    status_broker = "broker2:9092/topic2"
+    window.status_broker_edit.setText(status_broker)
+
+    file_broker = "broker3:9092/topic3"
+    window.command_widget.broker_line_edit.setText(file_broker)
+
+    use_start_time = True
+    window.command_widget.start_time_enabled.setChecked(use_start_time)
+
+    use_stop_time = True
+    window.command_widget.stop_time_enabled.setChecked(use_stop_time)
+
+    filename = "test.nxs"
+    window.command_widget.nexus_file_name_edit.setText(filename)
+
+    window._store_settings()
+
+    assert settings.value(FileWriterSettings.COMMAND_BROKER_ADDR) == command_broker
+    assert settings.value(FileWriterSettings.STATUS_BROKER_ADDR) == status_broker
+    assert settings.value(FileWriterSettings.FILE_BROKER_ADDR) == file_broker
+    assert (
+        extract_bool_from_qsettings(settings.value(FileWriterSettings.USE_START_TIME))
+        == use_start_time
+    )
+    assert (
+        extract_bool_from_qsettings(settings.value(FileWriterSettings.USE_STOP_TIME))
+        == use_stop_time
+    )
+    assert settings.value(FileWriterSettings.FILE_NAME) == filename
+
+
+def test_UI_stored_settings_are_shown_in_window(qtbot, instrument, settings):
+    command_broker = "broker:9092/topic2"
+    status_broker = "broker2:9092/topic3"
+    file_broker = "broker3:9092/topic4"
+    use_start_time = True
+    use_stop_time = False
+    filename = "test2.nxs"
+
+    settings.setValue(FileWriterSettings.STATUS_BROKER_ADDR, status_broker)
+    settings.setValue(FileWriterSettings.COMMAND_BROKER_ADDR, command_broker)
+    settings.setValue(FileWriterSettings.FILE_NAME, filename)
+    settings.setValue(FileWriterSettings.USE_START_TIME, use_start_time)
+    settings.setValue(FileWriterSettings.USE_STOP_TIME, use_stop_time)
+    settings.setValue(FileWriterSettings.FILE_BROKER_ADDR, file_broker)
+
+    # _restore_settings should be called on construction
+    window = FileWriterCtrl(instrument, settings)
+    qtbot.addWidget(window)
+
+    assert window.status_broker_edit.text() == status_broker
+    assert window.command_broker_edit.text() == command_broker
+    assert use_start_time == window.command_widget.start_time_enabled.isChecked()
+    assert use_stop_time == window.command_widget.stop_time_enabled.isChecked()
+    assert filename == window.command_widget.nexus_file_name_edit.text()
+    assert file_broker == window.command_widget.broker_line_edit.text()
+
+
+def test_UI_disable_stop_button_when_no_files_are_selected(qtbot, instrument, settings):
+    window = FileWriterCtrl(instrument, settings)
+    qtbot.addWidget(window)
+
+    assert not window.files_list.selectedIndexes()
+    assert not window.stop_file_writing_button.isEnabled()
